Fixed problem with TPC-C Delivery with Peloton. Added debug info for TPC-C StockLevel

+++ src/com/oltpbenchmark/benchmarks/tpcc/procedures/Delivery.java
@@ -206,7 +206,7 @@ public class Delivery extends TPCCProcedure {
             rs = null;
 
             int idx = 1; // HACK: So that we can debug this query
-            delivUpdateCustBalDelivCnt.setFloat(idx++, ol_total);
+            delivUpdateCustBalDelivCnt.setDouble(idx++, ol_total);
             delivUpdateCustBalDelivCnt.setInt(idx++, w_id);
             delivUpdateCustBalDelivCnt.setInt(idx++, d_id);
             delivUpdateCustBalDelivCnt.setInt(idx++, c_id);
